Changed mistake in reference to event.data.marker.cfMatrix in ARjs.MarkerControls Added missing reference to _this.data.minConfidence in arjs-anchor component
There was a mistake in ARjs.MarkerControls when event.data.marker.cfMatt was used to get the current confidence level when watching a marker using matrixcode. When logging the event data you can see that this property does not exist. The correct one is event.data.marker.cfMatrix.

There was also a missing reference to the property minConfidence in the schema of 'arjs-anchor' component. So when setting the minConfidence property in an a-frame entity, it was not registered and always defaulted to 0.6. The change I added now updates minConfidence to the value set in the component in the  entity:
e.g. <a-marker min-confidence="0.4" .... >